+++ app/src/androidTest/java/com/github/sdpcoachme/profile/ProfileActivityTest.kt
@@ -13,27 +13,27 @@ import com.github.sdpcoachme.CoachMeApplication
 import com.github.sdpcoachme.Dashboard.TestTags.Buttons.Companion.HAMBURGER_MENU
 import com.github.sdpcoachme.Dashboard.TestTags.Companion.BAR_TITLE
 import com.github.sdpcoachme.Dashboard.TestTags.Companion.DRAWER_HEADER
-import com.github.sdpcoachme.EditTextActivity.TestTags.Companion.TextFields.Companion.MAIN
 import com.github.sdpcoachme.R
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Buttons.Companion.MESSAGE_COACH
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.COACH_SWITCH
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.EMAIL
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.FIRST_NAME
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.LAST_NAME
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.LOCATION
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.PHONE
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.PROFILE_LABEL
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.SPORTS
 import com.github.sdpcoachme.data.UserInfoSamples.Companion.COACHES
 import com.github.sdpcoachme.data.UserInfoSamples.Companion.COACH_1
 import com.github.sdpcoachme.data.UserInfoSamples.Companion.COACH_2
 import com.github.sdpcoachme.data.UserInfoSamples.Companion.NON_COACHES
 import com.github.sdpcoachme.data.UserInfoSamples.Companion.NON_COACH_2
+import com.github.sdpcoachme.database.Database
 import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.Buttons.Companion.GO_TO_LOGIN_BUTTON
 import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.TextFields.Companion.ERROR_MESSAGE_FIELD
-import com.github.sdpcoachme.database.Database
 import com.github.sdpcoachme.location.autocomplete.MockLocationAutocompleteHandler.Companion.DEFAULT_LOCATION
 import com.github.sdpcoachme.messaging.ChatActivity
+import com.github.sdpcoachme.profile.EditTextActivity.TestTags.Companion.TextFields.Companion.MAIN
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Buttons.Companion.MESSAGE_COACH
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.COACH_SWITCH
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.EMAIL
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.FIRST_NAME
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.LAST_NAME
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.LOCATION
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.PHONE
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.PROFILE_LABEL
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.SPORTS
 import junit.framework.TestCase
 import org.hamcrest.CoreMatchers
 import org.junit.After
