+++ app/src/androidTest/java/com/github/sdpcoachme/profile/SelectSportsActivityTest.kt
@@ -16,20 +16,20 @@ import com.github.sdpcoachme.CoachMeApplication
 import com.github.sdpcoachme.Dashboard.TestTags.Buttons.Companion.HAMBURGER_MENU
 import com.github.sdpcoachme.Dashboard.TestTags.Companion.BAR_TITLE
 import com.github.sdpcoachme.Dashboard.TestTags.Companion.DRAWER_HEADER
+import com.github.sdpcoachme.auth.SignupActivity
+import com.github.sdpcoachme.data.Sports
+import com.github.sdpcoachme.data.UserInfo
+import com.github.sdpcoachme.data.UserLocationSamples.Companion.PARIS
+import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity
+import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.Buttons.Companion.GO_TO_LOGIN_BUTTON
+import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.TextFields.Companion.ERROR_MESSAGE_FIELD
+import com.github.sdpcoachme.location.MapActivity
 import com.github.sdpcoachme.profile.SelectSportsActivity.*
 import com.github.sdpcoachme.profile.SelectSportsActivity.TestTags.Buttons.*
 import com.github.sdpcoachme.profile.SelectSportsActivity.TestTags.Buttons.Companion.REGISTER
 import com.github.sdpcoachme.profile.SelectSportsActivity.TestTags.Companion.COLUMN
 import com.github.sdpcoachme.profile.SelectSportsActivity.TestTags.Companion.LIST_TITLE
 import com.github.sdpcoachme.profile.SelectSportsActivity.TestTags.MultiSelectListTag.Companion.ROW_TEXT_LIST
-import com.github.sdpcoachme.auth.SignupActivity
-import com.github.sdpcoachme.data.Sports
-import com.github.sdpcoachme.data.UserInfo
-import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity
-import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.Buttons.Companion.GO_TO_LOGIN_BUTTON
-import com.github.sdpcoachme.errorhandling.IntentExtrasErrorHandlerActivity.TestTags.TextFields.Companion.ERROR_MESSAGE_FIELD
-import com.github.sdpcoachme.data.UserLocationSamples.Companion.PARIS
-import com.github.sdpcoachme.map.MapActivity
 import junit.framework.TestCase
 import org.hamcrest.Matchers.*
 import org.junit.Before
