+++ app/src/main/java/com/github/sdpcoachme/location/MapActivity.kt
@@ -25,13 +25,13 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import com.github.sdpcoachme.CoachMeApplication
 import com.github.sdpcoachme.Dashboard
-import com.github.sdpcoachme.auth.ProfileActivity
 import com.github.sdpcoachme.data.UserInfo
-import com.github.sdpcoachme.errorhandling.ErrorHandlerLauncher
 import com.github.sdpcoachme.database.Database
-import com.github.sdpcoachme.map.MapActivity.TestTags.Companion.MAP
-import com.github.sdpcoachme.map.MapActivity.TestTags.Companion.MARKER
-import com.github.sdpcoachme.map.MapActivity.TestTags.Companion.MARKER_INFO_WINDOW
+import com.github.sdpcoachme.errorhandling.ErrorHandlerLauncher
+import com.github.sdpcoachme.location.MapActivity.TestTags.Companion.MAP
+import com.github.sdpcoachme.location.MapActivity.TestTags.Companion.MARKER
+import com.github.sdpcoachme.location.MapActivity.TestTags.Companion.MARKER_INFO_WINDOW
+import com.github.sdpcoachme.profile.ProfileActivity
 import com.github.sdpcoachme.ui.theme.CoachMeTheme
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
