+++ app/src/main/java/com/github/sdpcoachme/profile/ProfileActivity.kt
@@ -25,23 +25,24 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.github.sdpcoachme.*
+import com.github.sdpcoachme.CoachMeApplication
+import com.github.sdpcoachme.Dashboard
 import com.github.sdpcoachme.R
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Buttons.Companion.MESSAGE_COACH
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.COACH_SWITCH
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.EMAIL
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.FIRST_NAME
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.LAST_NAME
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.LOCATION
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.PHONE
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.PROFILE_LABEL
-import com.github.sdpcoachme.auth.ProfileActivity.TestTags.Companion.SPORTS
 import com.github.sdpcoachme.data.Sports
 import com.github.sdpcoachme.data.UserInfo
-import com.github.sdpcoachme.errorhandling.ErrorHandlerLauncher
 import com.github.sdpcoachme.database.Database
+import com.github.sdpcoachme.errorhandling.ErrorHandlerLauncher
 import com.github.sdpcoachme.location.autocomplete.LocationAutocompleteHandler
 import com.github.sdpcoachme.messaging.ChatActivity
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Buttons.Companion.MESSAGE_COACH
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.COACH_SWITCH
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.EMAIL
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.FIRST_NAME
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.LAST_NAME
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.LOCATION
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.PHONE
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.PROFILE_LABEL
+import com.github.sdpcoachme.profile.ProfileActivity.TestTags.Companion.SPORTS
 import com.github.sdpcoachme.ui.theme.CoachMeTheme
 import kotlinx.coroutines.future.await
 import java.util.concurrent.CompletableFuture
